THE PIGS GAME PROJECT
The pigs game project is a game deceloped using html css and javascript where two players playe the game. One is active player 0 and another one is active player 1. They roll the dice one by one. whatever value comes in dice that gets added to their current score.but if the value appeared on dice is 1 the current player looses all score. so it is necessary to hold the score also.BUt once you hold the score the dice rolled by another player and the game continues so on. Whiever player reaches 100 score first wins the game.